add mobilephonedata.py
