+++ src/transbigdata/traj.py
@@ -33,170 +33,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 import geopandas as gpd
 import pandas as pd
 import numpy as np
-from .grids import GPS_to_grid
-from .preprocess import id_reindex, clean_same
-
-
-def plot_activity(data, col=['stime', 'etime', 'LONCOL', 'LATCOL'],
-                  figsize=(10, 5), dpi=250):
-    '''
-    Plot the activity plot of individual
-
-    Parameters
-    ----------------
-    data : DataFrame
-        activity information of one person
-    col : List
-        The column name.[starttime,endtime,LONCOL,LATCOL] of activities
-    '''
-    stime, etime, LONCOL, LATCOL = col
-    activity = data.copy()
-    activity['date'] = activity[stime].dt.date
-    dates = list(activity['date'].astype(str).drop_duplicates())
-    dates_all = []
-    minday = min(dates)
-    maxday = max(dates)
-    import datetime
-    thisdate = minday
-    while thisdate != maxday:
-        dates_all.append(thisdate)
-        thisdate = str((pd.to_datetime(thisdate+' 00:00:00') +
-                       datetime.timedelta(days=1)).date())
-    dates = dates_all
-    import matplotlib.pyplot as plt
-    import numpy as np
-    activity['duration'] = (activity[etime]-activity[stime]).dt.total_seconds()
-    activity = activity[-activity['duration'].isnull()]
-    import time
-    activity['ststmp'] = activity[stime].astype(str).apply(
-        lambda x: time.mktime(
-            time.strptime(x, '%Y-%m-%d %H:%M:%S'))).astype('int64')
-    activity['etstmp'] = activity[etime].astype(str).apply(
-        lambda x: time.mktime(
-            time.strptime(x, '%Y-%m-%d %H:%M:%S'))).astype('int64')
-    activityinfo = activity[[LONCOL, LATCOL]].drop_duplicates()
-    indexs = list(range(1, len(activityinfo)+1))
-    np.random.shuffle(indexs)
-    activityinfo['index'] = indexs
-    import matplotlib as mpl
-    norm = mpl.colors.Normalize(vmin=0, vmax=len(activityinfo))
-    from matplotlib.colors import ListedColormap
-    import seaborn as sns
-    cmap = ListedColormap(sns.hls_palette(
-        n_colors=len(activityinfo), l=.5, s=0.8))
-    plt.figure(1, figsize, dpi)
-    ax = plt.subplot(111)
-    plt.sca(ax)
-    for day in range(len(dates)):
-        plt.bar(day, height=24*3600, bottom=0, width=0.4, color=(0, 0, 0, 0.1))
-        stime = dates[day]+' 00:00:00'
-        etime = dates[day]+' 23:59:59'
-        bars = activity[(activity['stime'] < etime) &
-                        (activity['etime'] > stime)].copy()
-        bars['ststmp'] = bars['ststmp'] - \
-            time.mktime(time.strptime(stime, '%Y-%m-%d %H:%M:%S'))
-        bars['etstmp'] = bars['etstmp'] - \
-            time.mktime(time.strptime(stime, '%Y-%m-%d %H:%M:%S'))
-        for row in range(len(bars)):
-            plt.bar(day,
-                    height=bars['etstmp'].iloc[row]-bars['ststmp'].iloc[row],
-                    bottom=bars['ststmp'].iloc[row],
-                    color=cmap(
-                        norm(
-                            activityinfo[
-                                (activityinfo[LONCOL] == bars[LONCOL].
-                                 iloc[row]) &
-                                (activityinfo[LATCOL] ==
-                                 bars[LATCOL].iloc[row])
-                            ]['index'].iloc[0])))
-    plt.xlim(-0.5, len(dates))
-    plt.ylim(0, 24*3600)
-    plt.xticks(range(len(dates)), [i[-5:] for i in dates])
-    plt.yticks(range(0, 24*3600+1, 3600),
-               pd.DataFrame({'t': range(0, 25)})['t'].astype('str')+':00')
-    plt.show()
-
-
-def traj_stay_move(data, params,
-                   col=['ID', 'dataTime', 'longitude', 'latitude'],
-                   activitytime=1800):
-    '''
-    Input trajectory data and gridding parameters, identify stay and move
-
-    Parameters
-    ----------------
-    data : DataFrame
-        trajectory data
-    params : List
-        gridding parameters
-    col : List
-        The column name, in the order of ['ID','dataTime','longitude',
-        'latitude']
-    activitytime : Number
-        How much time to regard as activity
-
-    Returns
-    ----------------
-    stay : DataFrame
-        stay information
-    move : DataFrame
-        move information
-    '''
-    uid, timecol, lon, lat = col
-    #Identify stay
-    stay = data.copy()
-    stay = stay.rename(columns={lon: 'lon', lat: 'lat',timecol:'stime'})
-    stay[timecol] = pd.to_datetime(stay[timecol])
-    stay['LONCOL'], stay['LATCOL'] = GPS_to_grid(
-        stay['lon'], stay['lat'], params)
-    #Number the status
-    stay['status_id'] = ((stay['LONCOL'] != stay['LONCOL'].shift()) |
-                        (stay['LATCOL'] != stay['LATCOL'].shift()) |
-                        (stay[uid] != stay[uid].shift())).astype(int)
-    stay['status_id'] = stay.groupby([uid])['status_id'].cumsum()
-    stay = stay.drop_duplicates(
-        subset=[uid, 'status_id'], keep='first').copy()
-    stay['etime'] = stay['stime'].shift(-1)
-    stay = stay[stay[uid] == stay[uid].shift(-1)].copy()
-    #Remove the duration shorter than given activitytime
-    stay['duration'] = (pd.to_datetime(stay['etime']) -
-                        pd.to_datetime(stay['stime'])).dt.total_seconds()
-    stay = stay[stay['duration'] >= activitytime].copy()
-    #Renumber the status
-    stay['status_id'] = ((stay['LONCOL'] != stay['LONCOL'].shift()) |
-                        (stay['LATCOL'] != stay['LATCOL'].shift()) |
-                        (stay[uid] != stay[uid].shift())).astype(int)
-    stay['status_id'] = stay.groupby([uid])['status_id'].cumsum()
-    stay_stime = stay.drop_duplicates(subset=[uid, 'status_id'], keep='first')[
-        [uid, 'stime', 'LONCOL', 'LATCOL', 'status_id']]
-    stay_etime = stay.drop_duplicates(subset=[uid, 'status_id'], keep='last')[
-        [uid, 'etime', 'LONCOL', 'LATCOL', 'status_id']]
-    stay = pd.merge(stay_stime, stay_etime, how='left').drop('status_id',axis = 1)
-    stay['lon'], stay['lat'] = grid_to_centre(
-        [stay['LONCOL'], stay['LATCOL']], params)
-    stay['duration'] = (pd.to_datetime(stay['etime']) -
-                        pd.to_datetime(stay['stime'])).dt.total_seconds()
-    #Identify move
-    move = stay.copy()
-    move['stime_next'] = move['stime'].shift(-1)
-    move['elon'] = move['lon'].shift(-1)
-    move['elat'] = move['lat'].shift(-1)
-    move['ELONCOL'] = move['LONCOL'].shift(-1)
-    move['ELATCOL'] = move['LATCOL'].shift(-1)
-    move[uid+'_next'] = move[uid].shift(-1)
-    move = move[move[uid+'_next'] == move[uid]
-                        ].drop(['stime', 'duration', uid+'_next'], axis=1)
-    move = move.rename(columns={'lon': 'slon',
-                                        'lat': 'slat',
-                                        'etime': 'stime',
-                                        'stime_next': 'etime',
-                                        'LONCOL': 'SLONCOL',
-                                        'LATCOL': 'SLATCOL',
-                                        })
-    move['duration'] = (
-        move['etime'] - move['stime']).dt.total_seconds()
-    return stay, move
-
+from .preprocess import id_reindex
 
 
 def traj_densify(data, col=['Vehicleid', 'Time', 'Lng', 'Lat'], timegap=15):
